Move locale types to types directory
