+++ config/locales.ts
@@ -1,4 +1,3 @@
-export type SupportedLocale = (typeof supportedLocales)[number];
-
+import type { SupportedLocale } from "@/types/locales";
 export const supportedLocales = ["bg", "en"] as const;
 export const defaultLocale: SupportedLocale = "bg";
