+++ translations/index.ts
@@ -1,6 +1,6 @@
 import bg from "./bg.json";
 import en from "./en.json";
-import type { SupportedLocale } from "@/config/locales";
+import type { SupportedLocale } from "@/types/locales";
 
 type TranslationObject = Record<string, string>;
 type TranslationsObject = Record<SupportedLocale, TranslationObject>;
