+++ types/locales.ts
@@ -1,0 +1,2 @@
+import { supportedLocales } from "@/config/locales";
+export type SupportedLocale = (typeof supportedLocales)[number];
